feat: create database migrations hook
